Move ROM::Repository to ROM::Gateway, add shim
Added a warning shim on inheritance for ROM::Repository
