+++ lib/rom/repository.rb
@@ -1,125 +1,19 @@
+require 'rom/gateway'
+
 module ROM
   # Abstract repository class
   #
+  # This is a transitional placeholder, deprecating the Repository class.
+  #
   # @api public
-  class Repository
-    # Return connection object
-    #
-    # @return [Object] type varies depending on the repository
-    #
-    # @api public
-    attr_reader :connection
+  class Repository < Gateway
+    def self.inherited(_klass)
+      warn <<-MSG.gsub(/^\s+/, '')
+        Inheriting from ROM::Repository is deprecated and will be removed in 1.0.0.
+        Please inherit from ROM::Gateway instead.
 
-    # Setup a repository
-    #
-    # @overload setup(type, *args)
-    #   Sets up a single-repository given a repository type.
-    #   For custom repositories, create an instance and pass it directly.
-    #
-    #   @param [Symbol] type
-    #   @param [Array] *args
-    #
-    # @overload setup(repository)
-    #   @param [Repository] repository
-    #
-    # @return [Repository] a specific repository subclass
-    #
-    # @example
-    #   module SuperDB
-    #     class Repository < ROM::Repository
-    #       def initialize(options)
-    #       end
-    #     end
-    #   end
-    #
-    #   ROM.register_adapter(:super_db, SuperDB)
-    #
-    #   Repository.setup(:super_db, some: 'options')
-    #   # SuperDB::Repository.new(some: 'options') is called
-    #
-    #   # or alternatively
-    #   super_db = Repository.setup(SuperDB::Repository.new(some: 'options'))
-    #   Repository.setup(super_db)
-    #
-    # @api public
-    def self.setup(repository_or_scheme, *args)
-      case repository_or_scheme
-      when String
-        raise ArgumentError, <<-STRING.gsub(/^ {10}/, '')
-          URIs without an explicit scheme are not supported anymore.
-          See https://github.com/rom-rb/rom/blob/master/CHANGELOG.md
-        STRING
-      when Symbol
-        class_from_symbol(repository_or_scheme).new(*args)
-      else
-        if args.empty?
-          repository_or_scheme
-        else
-          raise ArgumentError, "Can't accept arguments when passing an instance"
-        end
-      end
-    end
-
-    # Get repository subclass for a specific adapter
-    #
-    # @param [Symbol] type adapter identifier
-    #
-    # @return [Class]
-    #
-    # @api private
-    def self.class_from_symbol(type)
-      begin
-        require "rom/#{type}"
-      rescue LoadError
-        raise AdapterLoadError, "Failed to load adapter rom/#{type}"
-      end
-
-      adapter = ROM.adapters.fetch(type)
-      adapter.const_get(:Repository)
-    end
-
-    # A generic interface for setting up a logger
-    #
-    # @api public
-    def use_logger(*)
-      # noop
-    end
-
-    # A generic interface for returning default logger
-    #
-    # @api public
-    def logger
-      # noop
-    end
-
-    # Extension hook for adding repository-specific behavior to a command class
-    #
-    # @param [Class] klass command class
-    # @param [Object] _dataset dataset that will be used with this command class
-    #
-    # @return [Class]
-    #
-    # @api public
-    def extend_command_class(klass, _dataset)
-      klass
-    end
-
-    # Schema inference hook
-    #
-    # Every repository that supports schema inference should implement this method
-    #
-    # @return [Array] array with datasets and their names
-    #
-    # @api private
-    def schema
-      []
-    end
-
-    # Disconnect is optional and it's a no-op by default
-    #
-    # @api public
-    def disconnect
-      # noop
+        #{caller.detect { |l| !l.include?('lib/rom') }}
+      MSG
     end
   end
 end
